+++ src/app/Components/ComponentsSamples/BottomSheets/Sheets/NavigationPage2.xaml.cs
@@ -1,0 +1,25 @@
+using DIPS.Mobile.UI.Components.BottomSheets;
+
+namespace Components.ComponentsSamples.BottomSheets.Sheets;
+
+public partial class NavigationPage2
+{
+    private readonly BottomSheet m_sheet;
+
+    public NavigationPage2(BottomSheet sheet)
+    {
+        m_sheet = sheet;
+        InitializeComponent();
+    }
+
+    private async void OnGoDeeperTapped(object? sender, EventArgs e)
+    {
+        await m_sheet.PushAsync(new NavigationPage3(m_sheet));
+    }
+
+    private async void OnPopToRootClicked(object? sender, EventArgs e)
+    {
+        await m_sheet.PopToRootAsync();
+    }
+}
+