+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [60.3.0]
+- [Bottomsheet] Added page navigation inside bottomsheet
+- [Checkmark] Added `AddStrokeToCheckmark` property to display a visible stroke around the checkmark
+- [CheckmarkListItem] Added `AddStrokeToCheckmark` property to display a visible stroke around the checkmark icon
+
 ## [60.2.15]
 - [ImageCapture][iOS] Fixed full-screen gallery previews showing black images when opened from thumbnails.